+++ src/contexts/Accounts/Balances/defaults.ts
@@ -4,7 +4,6 @@
 import BigNumber from 'bignumber.js';
 import {
   Balance,
-  BalanceLedger,
   BalancesContextInterface,
   Nominations,
 } from 'contexts/Accounts/Balances/types';
@@ -15,14 +14,6 @@ export const balance: Balance = {
   miscFrozen: new BigNumber(0),
   feeFrozen: new BigNumber(0),
   freeAfterReserve: new BigNumber(0),
-};
-
-export const ledger: BalanceLedger = {
-  address: null,
-  stash: null,
-  active: new BigNumber(0),
-  total: new BigNumber(0),
-  unlocking: [],
 };
 
 export const nominations: Nominations = {
@@ -36,10 +27,6 @@ export const defaultBalancesContext: BalancesContextInterface = {
   // eslint-disable-next-line
   getAccountBalance: (address) => balance,
   // eslint-disable-next-line
-  getLedgerForStash: (address) => ledger,
-  // eslint-disable-next-line
-  getLedgerForController: (address) => null,
-  // eslint-disable-next-line
   getAccountLocks: (address) => [],
   // eslint-disable-next-line
   getBondedAccount: (address) => null,
@@ -47,7 +34,6 @@ export const defaultBalancesContext: BalancesContextInterface = {
   getAccountNominations: (address) => [],
   // eslint-disable-next-line
   isController: (address) => false,
-  accounts: [],
+  balancesAccounts: [],
   existentialAmount: new BigNumber(0),
-  ledgers: [],
 };
