+++ src/contexts/Accounts/Balances/index.tsx
@@ -3,7 +3,6 @@
 
 import BigNumber from 'bignumber.js';
 import {
-  BalanceLedger,
   BalancesAccount,
   BalancesContextInterface,
 } from 'contexts/Accounts/Balances/types';
@@ -27,59 +26,47 @@ export const BalancesProvider = ({
   children: React.ReactNode;
 }) => {
   const { api, isReady, network, consts } = useApi();
-  const { accounts: connectAccounts, addExternalAccount } = useConnect();
+  const { accounts, addExternalAccount } = useConnect();
 
   // existential amount of unit for an account
   const existentialAmount = consts.existentialDeposit;
 
   // balance accounts state
-  const [accounts, setAccounts] = useState<Array<BalancesAccount>>([]);
-  const accountsRef = useRef(accounts);
+  const [balancesAccounts, setBalancesAccounts] = useState<
+    Array<BalancesAccount>
+  >([]);
+  const balancesAccountsRef = useRef(balancesAccounts);
 
   // balance subscriptions state
   const [unsubsBalances, setUnsubsBalances] = useState<AnyApi>([]);
   const unsubsBalancesRef = useRef<AnyApi>(unsubsBalances);
 
-  // account ledgers to separate storage
-  const [ledgers, setLedgers] = useState<Array<BalanceLedger>>([]);
-  const ledgersRef = useRef(ledgers);
-
-  // ledger subscriptions state
-  const [unsubsLedgers, setUnsubsLedgers] = useState<AnyApi>([]);
-  const unsubsLedgersRef = useRef<AnyApi>(unsubsLedgers);
-
   // fetch account balances & ledgers. Remove or add subscriptions
   useEffect(() => {
     if (isReady) {
       // local updated values
-      let newAccounts = accountsRef.current;
-      let newLedgers = ledgersRef.current;
+      let newAccounts = balancesAccountsRef.current;
       const newUnsubsBalances = unsubsBalancesRef.current;
-      const newUnsubsLedgers = unsubsLedgersRef.current;
 
       // get accounts removed: use these to unsubscribe
-      const accountsRemoved = accountsRef.current.filter(
+      const accountsRemoved = balancesAccountsRef.current.filter(
         (a: BalancesAccount) =>
-          !connectAccounts.find((c: ImportedAccount) => c.address === a.address)
+          !accounts.find((c: ImportedAccount) => c.address === a.address)
       );
       // get accounts added: use these to subscribe
-      const accountsAdded = connectAccounts.filter(
+      const accountsAdded = accounts.filter(
         (c: ImportedAccount) =>
-          !accountsRef.current.find(
+          !balancesAccountsRef.current.find(
             (a: BalancesAccount) => a.address === c.address
           )
       );
       // update accounts state for removal
-      newAccounts = accountsRef.current.filter((a: BalancesAccount) =>
-        connectAccounts.find((c: ImportedAccount) => c.address === a.address)
-      );
-      // update ledgers state for removal
-      newLedgers = ledgersRef.current.filter((l: BalanceLedger) =>
-        connectAccounts.find((c: ImportedAccount) => c.address === l.address)
+      newAccounts = balancesAccountsRef.current.filter((a: BalancesAccount) =>
+        accounts.find((c: ImportedAccount) => c.address === a.address)
       );
 
       // update accounts state and unsubscribe if accounts have been removed
-      if (newAccounts.length < accountsRef.current.length) {
+      if (newAccounts.length < balancesAccountsRef.current.length) {
         // unsubscribe from removed balances
         accountsRemoved.forEach((a: BalancesAccount) => {
           const unsub = unsubsBalancesRef.current.find(
@@ -97,25 +84,7 @@ export const BalancesProvider = ({
           setUnsubsBalances,
           unsubsBalancesRef
         );
-        setStateWithRef(newAccounts, setAccounts, accountsRef);
-      }
-
-      // update ledgers state and unsubscribe if accounts have been removed
-      if (newLedgers.length < ledgersRef.current.length) {
-        // unsubscribe from removed ledgers if it exists
-        accountsRemoved.forEach((a: BalancesAccount) => {
-          const unsub = unsubsLedgersRef.current.find(
-            (u: AnyApi) => u.key === a.address
-          );
-          if (unsub) {
-            unsub.unsub();
-            // remove unsub from balances
-            newUnsubsLedgers.filter((u: AnyApi) => u.key !== a.address);
-          }
-        });
-        // commit state updates
-        setStateWithRef(newUnsubsLedgers, setUnsubsLedgers, unsubsLedgersRef);
-        setStateWithRef(newLedgers, setLedgers, ledgersRef);
+        setStateWithRef(newAccounts, setBalancesAccounts, balancesAccountsRef);
       }
 
       // if accounts have changed, update state with new unsubs / accounts
@@ -124,7 +93,7 @@ export const BalancesProvider = ({
         handleSubscribe(accountsAdded);
       }
     }
-  }, [connectAccounts, network, isReady]);
+  }, [accounts, network, isReady]);
 
   // unsubscribe from everything on unmount
   useEffect(() => {
@@ -139,20 +108,13 @@ export const BalancesProvider = ({
     Promise.all(
       accountsAdded.map((a: ImportedAccount) => subscribeToBalances(a.address))
     );
-    // subscribe to ledgers
-    Promise.all(
-      accountsAdded.map((a: ImportedAccount) => subscribeToLedger(a.address))
-    );
   };
 
   /*
-   * Unsubscrbe all balance subscriptions
+   * Unsubscrbe balance subscriptions
    */
   const unsubscribeAll = () => {
     Object.values(unsubsBalancesRef.current).forEach(({ unsub }: AnyApi) => {
-      unsub();
-    });
-    Object.values(unsubsLedgersRef.current).forEach(({ unsub }: AnyApi) => {
       unsub();
     });
   };
@@ -206,9 +168,8 @@ export const BalancesProvider = ({
         // add bonded (controller) account as external account if not presently imported
         if (_bonded) {
           if (
-            connectAccounts.find(
-              (s: ImportedAccount) => s.address === _bonded
-            ) === undefined
+            accounts.find((s: ImportedAccount) => s.address === _bonded) ===
+            undefined
           ) {
             addExternalAccount(_bonded, 'system');
           }
@@ -225,11 +186,11 @@ export const BalancesProvider = ({
               };
 
         // remove stale account if it's already in list.
-        const newAccounts = Object.values(accountsRef.current)
+        const newAccounts = Object.values(balancesAccountsRef.current)
           .filter((a: BalancesAccount) => a.address !== address)
           .concat(newAccount);
 
-        setStateWithRef(newAccounts, setAccounts, accountsRef);
+        setStateWithRef(newAccounts, setBalancesAccounts, balancesAccountsRef);
       }
     );
 
@@ -244,76 +205,9 @@ export const BalancesProvider = ({
     return unsub;
   };
 
-  const subscribeToLedger = async (address: string) => {
-    if (!api) return;
-
-    const unsub: () => void = await api.queryMulti<AnyApi>(
-      [[api.query.staking.ledger, address]],
-      async ([l]): Promise<void> => {
-        let ledger: BalanceLedger;
-
-        const _ledger = l.unwrapOr(null);
-        // fallback to default ledger if not present
-        if (_ledger !== null) {
-          const { stash, total, active, unlocking } = _ledger;
-
-          // format unlocking chunks
-          const _unlocking = [];
-          for (const u of unlocking.toHuman()) {
-            const { era, value } = u;
-
-            _unlocking.push({
-              era: Number(rmCommas(era)),
-              value: new BigNumber(rmCommas(value)),
-            });
-          }
-
-          // add stash as external account if not present
-          if (
-            connectAccounts.find(
-              (s: ImportedAccount) => s.address === stash.toHuman()
-            ) === undefined
-          ) {
-            addExternalAccount(stash.toHuman(), 'system');
-          }
-
-          ledger = {
-            address,
-            stash: stash.toHuman(),
-            active: new BigNumber(rmCommas(active.toString())),
-            total: new BigNumber(rmCommas(total.toString())),
-            unlocking: _unlocking,
-          };
-
-          // remove stale account if it's already in list, and concat.
-          let _ledgers = Object.values(ledgersRef.current);
-          _ledgers = _ledgers
-            .filter((_l: BalanceLedger) => _l.stash !== ledger.stash)
-            .concat(ledger);
-
-          setStateWithRef(_ledgers, setLedgers, ledgersRef);
-        } else {
-          // no ledger: remove stale account if it's already in list.
-          let _ledgers = Object.values(ledgersRef.current);
-          _ledgers = _ledgers.filter(
-            (_l: BalanceLedger) => _l.address !== address
-          );
-          setStateWithRef(_ledgers, setLedgers, ledgersRef);
-        }
-      }
-    );
-
-    const _unsubs = unsubsLedgersRef.current.concat({
-      key: address,
-      unsub,
-    });
-    setStateWithRef(_unsubs, setUnsubsLedgers, unsubsLedgersRef);
-    return unsub;
-  };
-
   // get an account's balance metadata
   const getAccountBalance = (address: MaybeAccount) => {
-    const account = accountsRef.current.find(
+    const account = balancesAccountsRef.current.find(
       (a: BalancesAccount) => a.address === address
     );
     if (account === undefined) {
@@ -326,38 +220,9 @@ export const BalancesProvider = ({
     return balance;
   };
 
-  // get a stash account's ledger metadata
-  const getLedgerForStash = (address: MaybeAccount) => {
-    const ledger = ledgersRef.current.find(
-      (l: BalanceLedger) => l.stash === address
-    );
-    if (ledger === undefined) {
-      return defaults.ledger;
-    }
-    if (ledger.stash === undefined) {
-      return defaults.ledger;
-    }
-    return ledger;
-  };
-
-  // get a controler account's ledger
-  // returns null if ledger does not exist.
-  const getLedgerForController = (address: MaybeAccount) => {
-    const ledger = ledgersRef.current.find(
-      (l: BalanceLedger) => l.address === address
-    );
-    if (ledger === undefined) {
-      return null;
-    }
-    if (ledger.address === undefined) {
-      return null;
-    }
-    return ledger;
-  };
-
   // get an account's locks metadata
   const getAccountLocks = (address: MaybeAccount) => {
-    const account = accountsRef.current.find(
+    const account = balancesAccountsRef.current.find(
       (a: BalancesAccount) => a.address === address
     );
     if (account === undefined) {
@@ -370,7 +235,7 @@ export const BalancesProvider = ({
 
   // get an account's bonded (controller) account)
   const getBondedAccount = (address: MaybeAccount) => {
-    const account = accountsRef.current.find(
+    const account = balancesAccountsRef.current.find(
       (a: BalancesAccount) => a.address === address
     );
     if (account === undefined) {
@@ -382,7 +247,7 @@ export const BalancesProvider = ({
 
   // get an account's nominations
   const getAccountNominations = (address: MaybeAccount) => {
-    const account = accountsRef.current.find(
+    const account = balancesAccountsRef.current.find(
       (a: BalancesAccount) => a.address === address
     );
     if (account === undefined) {
@@ -399,7 +264,7 @@ export const BalancesProvider = ({
 
   // get an account
   const getAccount = (address: MaybeAccount) => {
-    const account = accountsRef.current.find(
+    const account = balancesAccountsRef.current.find(
       (a: BalancesAccount) => a.address === address
     );
     if (account === undefined) {
@@ -410,7 +275,7 @@ export const BalancesProvider = ({
 
   // check if an account is a controller account
   const isController = (address: MaybeAccount) => {
-    const existsAsController = accountsRef.current.filter(
+    const existsAsController = balancesAccountsRef.current.filter(
       (a: BalancesAccount) => (a?.bonded || '') === address
     );
     return existsAsController.length > 0;
@@ -421,15 +286,12 @@ export const BalancesProvider = ({
       value={{
         getAccount,
         getAccountBalance,
-        getLedgerForStash,
-        getLedgerForController,
         getAccountLocks,
         getBondedAccount,
         getAccountNominations,
         isController,
         existentialAmount,
-        accounts: accountsRef.current,
-        ledgers: ledgersRef.current,
+        balancesAccounts: balancesAccountsRef.current,
       }}
     >
       {children}
