+++ src/contexts/Accounts/Balances/types.ts
@@ -2,20 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import BigNumber from 'bignumber.js';
-import { AnyApi, MaybeAccount } from 'types';
-
-export interface UnlockChunk {
-  era: number;
-  value: BigNumber;
-}
-
-export interface BalanceLedger {
-  address: MaybeAccount;
-  stash: string | null;
-  active: BigNumber;
-  total: BigNumber;
-  unlocking: Array<UnlockChunk>;
-}
+import { MaybeAccount } from 'types';
 
 export interface BondedAccount {
   address: string;
@@ -39,7 +26,7 @@ export interface BalancesAccount {
   address?: string;
   balance?: Balance;
   bonded?: string;
-  ledger?: BalanceLedger;
+  // ledger?: BalanceLedger;
   locks?: Array<Lock>;
   nominations?: Nominations;
 }
@@ -54,13 +41,10 @@ export type Targets = string[];
 export interface BalancesContextInterface {
   getAccount: (address: MaybeAccount) => BalancesAccount | null;
   getAccountBalance: (address: MaybeAccount) => Balance;
-  getLedgerForStash: (address: MaybeAccount) => BalanceLedger;
-  getLedgerForController: (address: MaybeAccount) => BalanceLedger | null;
   getAccountLocks: (address: MaybeAccount) => Array<Lock>;
   getBondedAccount: (address: MaybeAccount) => string | null;
   getAccountNominations: (address: MaybeAccount) => Targets;
   isController: (address: MaybeAccount) => boolean;
-  accounts: Array<BalancesAccount>;
+  balancesAccounts: Array<BalancesAccount>;
   existentialAmount: BigNumber;
-  ledgers: AnyApi;
 }
